Object registry
Instances and Enums have the property where all references to them in
Luau have the same userdata. This is the groundwork for that property.

One way to check userdata equality is to index a table. If two different
expressions refer to the same value when used as the table index, then
they are the same userdata (as tables use raw equality).